Remove General settings section
