+++ app/src/main/java/eu/kanade/presentation/more/settings/screen/SettingsSearchScreen.kt
@@ -284,7 +284,6 @@ private fun getLocalizedBreadcrumb(path: String, node: String?): String {
 }
 
 private val settingScreens = listOf(
-    SettingsGeneralScreen,
     SettingsAppearanceScreen,
     SettingsLibraryScreen,
     SettingsReaderScreen,
